refactor: WebP 고화질 png 변환
